refactor: extract alert normalization and evaluation
